requirements.txt
dnspython==1.13.0
python-etcd==0.4.3
urllib3==1.15.1
